Remove RelayQL_GENERATED
Summary:
`RelayQL_GENERATED` served as a marker when we had 2 babel plugins stacked to
tell the classic compiler to disable some validation. We now have a single
plugin and can communicate the `disableValidation` flag directly. Instead of
getting the `RelayQL` helpers from a require, we instead pass them in as an
argument into the generated function.

Reviewed By: leebyron

Differential Revision: D5257109

fbshipit-source-id: 74f6dae49ca95aeb02695d20cd5cff5c8337db3c