Reapply: remove soft error mode
Summary:
This removes a "soft error" mode for the Relay Babel plugin where instead of
failing to transform code, the plugin would generate a `throw new Error(...)`
statement. This turns out to be a footgun since it removes the compiler safety
from guarding agains shipping code that's invalid according to the schema.

Reviewed By: leebyron

Differential Revision: D5273097

fbshipit-source-id: a59b737a0984aa2e0f67325d31a4f8716035bd91