Better error reporting
Summary:
Introducing a `RelayReporter` interface to abstract logging. For now, we only
use it for error reporting, but this could potentially be expanded to any
logging we do.
This could make it easier to log to a backend system or have a JSON reporter
that logs in a machine consumable format.

Error have a `location` to allow us to see where the error was thrown more
easily than going through a stack trace after babelification.

Reviewed By: leebyron

Differential Revision: D5308276

fbshipit-source-id: 693fb79e7c145f2ee42c03a3374a253294fa2353