Fix bug where RelayModernFragmentSpecResolver was not resolving data correctly
Summary:
Fixes #1750
Fixes #1904

The RelayModernFragmentSpecResolver was not resolving data correctly when the item at end of the array passed in props was removed.

i.e. given that my fragment is resolving data for `[zuck, beast]`, when this changes to `[zuck]`, the resolved data should reflect this. Previously it was returning the existing data as it was not checking if the length of the resolvers was not equal to the previous data length.

Does this need to be fixed in classic too?
Closes https://github.com/facebook/relay/pull/1907

Reviewed By: yuzhi

Differential Revision: D5321129

Pulled By: kassens

fbshipit-source-id: 78a687c47456b5eba9f88d77a944bbb4eff3d1eb