build should return with error code in case of errors
Summary:
In validation mode, the compiler correctly returned an error code for either an
error or unclean build. That's why we returned `true` in case of caught errors
for "has changes".

A simple build on the other hand shouldn't return with an error code if
everything went well, but there were changes.

This is the reason the boolean return type for the result of the build in
different places wasn't enough and this diff introduces an enum indicating
either a clean status, changes, or an error.

Reviewed By: JenniferWang

Differential Revision: D5365472

fbshipit-source-id: 3e0a64241d6ddc8dd508b1b0c3bc59c7bd1a25ce