add include/exclude options to compiler
Summary:
It would be nice to be a little more flexible with the search paths for `relay-compiler`. As of https://github.com/facebook/relay/pull/1686 `relay-compiler` ignores node_modules. We have several Relay apps that share common components through node_modules and need a way to tell the compiler which ones to search. This change will let us add several more specific search paths and override the setting to ignore node_modules.

I was careful to do this in a way that is not a breaking change.
Closes https://github.com/facebook/relay/pull/1921

Reviewed By: JenniferWang

Differential Revision: D5410197

Pulled By: kassens

fbshipit-source-id: e4fd90cb57be985458342ee9b35972982d5c462b