better warning when Codegen#deleteExtraFiles fails
Summary:
Codegen#deleteExtraFiles uses unlink to delete extra files in the generated directory, but if it can't delete something, it has a fairly cryptic error message (unlink is not something most peopl relat to delete).

This is just an easy way to make the error message better without using lstat.

Reviewed By: kassens

Differential Revision:
D5422105

Tags: relay

fbshipit-source-id: c465a8c4e4e3e69974d2ad6b12b7ce225cae0e51