Change readyState to getDefaultState value after retry
As exemplified in the QueryRenderer documentation [0], the render method
receives error and props as arguments, and the component is supposed to show an
error screen, a loading screen or the component based on those properties.

If error isn't null, show the error; if props isn't null, show the component;
otherwise, show the loading.

However, after calling the retry() function, the error isn't set to null. This
way, it can't know if it is loading or the same error occurred again.

With this patch, retry() changes the value of readyState to the same as the
initial value: all values set to null, telling the component that it is loading.

[0]: https://github.com/facebook/relay/blob/b4b086a8ce0005d1d13af4a96112d5c6d87f3e4c/docs/modern/QueryRenderer.md#L33-L40