Add requisite fields for @connection directives on fragments
Summary: Currently, when using `createPaginationContainer`, we require you to explicitly ask for `pageInfo` even if you don't need to access the data. This is counter-intuitive to the idea of only requesting the data you need. To get rid of that, we're generating those fields automatically on fragments if they weren't explicitly asked for to be used by relay for the pagination container.

Reviewed By: yuzhi

Differential Revision: D5440802

fbshipit-source-id: aac4f4a9f84041db82c1ade0e6712657cd87be5c