add a setting to configure flow types of custom scalars
Summary:
This adds a config that can contain a mapping from custom GraphQL scalars to
existing scalars, for example: `{URL: 'String'}` would map a `URL` custom scalar
to a the same flow type as `String`.

This is not work on adding support for (de)-serializing from custom JS types,
such as `URL`.

Reviewed By: leebyron

Differential Revision: D5627134

fbshipit-source-id: 90e94fb3bb45bc6856fe445053689165e7ad6db9