Add support for NullValue ast kind to RelayParser._transformValue()
Summary:
Hi there! This patch is the result of debugging gatsbyjs/gatsby#1689. It seems like the switch statement inside of `RelayParser._transformValue()` was missing a case for `NullValue`, so queries which contained a `null` literal would throw an error.

Adding a case fixes the downstream issue that some folks have experienced with Gatsby, but I'm not sure if there's a better way to solve this problem.
Closes https://github.com/facebook/relay/pull/2035

Reviewed By: leebyron

Differential Revision: D5650583

Pulled By: kassens

fbshipit-source-id: 9e76f553b35ab106711930150045d285670d36e2