make RelayConnection injectable
Summary:
For legacy reasons, we need the connection fields to be configurable. This
changes from a hacky module override to an exposed API that simplifies the
setup.

For Relay Modern, the impact should be minimal as the hottest code path this
impacts is probably the connection handler which only runs after normilization
on fields marked with `connection`.

For Classic mode, there's more potential impact as the config is read during
more tree traversal steps. By now, this seems okay as classic usage should go
down more and more over time.

Reviewed By: leebyron

Differential Revision: D5691273

fbshipit-source-id: 814172e1c4f6a2468aa8e80c729641521773fc93