Permit bidirectional PageInfo
The existing specification for PageInfo requires that `hasPreviousPage`
be false when paginating forward, which is quite limiting. Adjust that
part of the specification to permit setting `hasPreviousPage` when
paginating forward, and symmetrically permit setting `hasNextPage` when
paginating backward. Implementations may still hard-code false if
calculating the correct value would be inefficient.

Making this change exposed an inconsistency in handling unusual cases
such as when a client provides only `last` and `after` arguments. I
adjusted a few other parts of the algorithm to hopefully clarify how
best to handle these cases.