Move parts of CONTRIBUTING to ISSUE_TEMPLATE
Summary:
I noticed that a lot of questions that are better suited for SO are coming in here. So maybe by using GitHub's `ISSUE_TEMPLATE` we can circumvent that.

Also I found it a bit tedious that one can't simply create a jsfiddle to recreate a problem with Relay, so I changed `relay-starter-kit` a bit to make it work with Glitch and included a link to it in the `ISSUE_TEMPLATE`.

Lmkwyt :)
Closes https://github.com/facebook/relay/pull/2062

Differential Revision: D5789330

Pulled By: leebyron

fbshipit-source-id: 958bd285f9d4f773d2d3d38305d309e3c77d7a8b