deprecate graphql.experimental
Summary:
The only purpose for `graphql.experimental` was to enable `arguments`. It's
not really something tied to the compat layer and often used in Modern only
apps. If we decide to discourage `arguments`, we could change the name of the
directives.

In this step `graphql.experimental` is now just an alias for `graphql`, next step is adding a deprecation warning.

Reviewed By: leebyron

Differential Revision: D5791996

fbshipit-source-id: 485f9bf7dc4ef7b60e994882cfcdb1eb722da00b