Add documentation for granular mutation updates (via optimisticUpdater and updater) (fixes #1652)
Summary:
The purpose of this PR is to add much needed document for Relay Modern. The documentation I have added adds some clarity on how to transverse and mutate the Relay store for advanced mutations. This is necessary for performing equilateral updates optimistically and in response to a mutation. Many developers on Reactiflux and StackOverflow are seeking this information and it would be valuable to have it more readily accessible.

I tried to keep it as low level as possible and relevant. It is my belief that some of the docs have been designed to be more entry and mid level so I've indicated that these APIs are more advanced.

Right now there is a link inside the documentation to the following API
https://github.com/facebook/relay/blob/d0310d69012bba615dacf614319bcf47ee2a0f3f/packages/relay-runtime/ARCHITECTURE.md

If this is appropriate, no further action is necessary, though I believe this document should be exposed, and maybe the documentation I have added should be co-located with Architecture.md.

Since I have only modified markdown files, testing is not relevant but here are the results from `npm test`.
```
Test Suites 178 passed, 178 total
Tests       2118 passed, 2118 total
Snapshots   12 passed, 12 total
Time        26.834s
Ran all test suites.
```

Lastly I have added an example that's based on the TodoMVC basic example. Please let me know that it is idiomatic or needs reworking.

Thanks! It's been a pleasure working with Relay Modern.
Closes https://github.com/facebook/relay/pull/2014

Differential Revision: D5789674

Pulled By: leebyron

fbshipit-source-id: f87191411d178210533b336cbc73ab4d25e8eb08