Flatten fragments more aggressively fixing type propagation issue
Summary: While looking into Flow errors in the previously-unchecked `FlattenTransform` I stumbled across this issue. Where we were accessing `type` on a `Condition` node, which is a non-existent property and so always `undefined`. My reading of this is that it should actually be accessing `type` from the `FlattenState` instead, and with that change, we eliminate one Flow error and also guarantee that we always have `type` info in the `shouldFlattenInlineFragment` check; as a result, with this diff we end up flattening more fragments than we did before. Visual inspection of the changed outputs looks correct to me.

Reviewed By: kassens

Differential Revision: D5821122

fbshipit-source-id: 3bb65fd0c8b7f428dcff2439bfd5e1c77a400914