ReactRelayQueryRenderer: Change readyState to getDefaultState value after retry
Summary:
As exemplified in the [QueryRenderer documentation](https://github.com/facebook/relay/blob/b4b086a8ce0005d1d13af4a96112d5c6d87f3e4c/docs/modern/QueryRenderer.md#L33-L40), the `render` method receives `error` and `props` as arguments, and the component is supposed to show an error screen, a loading screen or the component based on those properties.

If `error` isn't `null`, show the error; if `props` isn't `null`, show the component; otherwise, show the loading.

However, after calling the `retry()` function, the `error` argument isn't passed again as `null`. This way, the component can't know if it is loading or the same error has just occurred again.

With this patch, `retry()` changes the value of `readyState` to the same as the initial value: all values set to `null`, telling the component that it is loading.
Closes https://github.com/facebook/relay/pull/2015

Reviewed By: kassens

Differential Revision: D5811379

Pulled By: leebyron

fbshipit-source-id: 3d45b8e32d4aa2d5841a824811665e9bf3003707