Add prettier config file
Summary:
Prettier 1.6 now supports config files, makes it a lot easier when working with multiple projects with different configs. 1.7 also adds pragma support so only files with format are formatted.

We could eventually remove parameters from the prettier script but that would also require updating the version of prettier to 1.6+ and don't want to do that in this PR since it will change a bunch of files if some formating changed in those newer versions.

Used the dotfile since I don't mind a million dotfiles and rather not add stuff we almost never care about in package.json.
Closes https://github.com/facebook/relay/pull/2123

Reviewed By: leebyron

Differential Revision: D5905794

Pulled By: kassens

fbshipit-source-id: 315625402b3c0c2ac69c5573468151f0ba5a403c