split out InlineFragmentsTransform from FlattenTransform
Summary: The `FlattenTransform` is pretty crazy in all the things it did with multiple options. This splits out the inlining and removing of fragments as well as the `relay(mask: false)` inlining splitting out the Relay specific behavior from the generic `graphql-compiler`. It still has to be moved out of the `graphql-compiler` directory as a next step.

Reviewed By: wincent

Differential Revision: D5912666

fbshipit-source-id: 956e46957af4dad5658264e9c41f566deec69ce8