dedupe JSON in generated files
Summary:
This uses the dedupeJSON utility on the generated JSON files. This reduces some
files to about 10% of the original minified, non-compressed size and about the
same compressed size.

Downside is that it makes the generated files less readable, but I guess they're
a bit too large to manually look at anyway.

Reviewed By: leebyron

Differential Revision: D5926638

fbshipit-source-id: fb852fc504f3147e352d1f6de2059c0fb348cdac