support fragments with only client fields
Summary:
The general `IRTransformer` should not strip fragments or roots that have no leftover selections. If the fragment is used anywhere (or someone expects a query in the result), it should be explicitly be removed. Otherwise, we cannot tell the difference between a missing fragment and an implicitly removed empty fragment.

This fixes an issue with fragments that only contain client extension fields.

Reviewed By: leebyron

Differential Revision: D6209740

fbshipit-source-id: d16e29d9b80575262e3a34856339525be5420c4d