generate hash on modern generated files
Summary:
This adds a hash of the normalized (via `graphql.print`) source text of fragments and operations to the generated files. This will be used together with the folowing diff to validate that the Relay build has run and help people detecting this non-obvious error case.

This is not ideal, as a fragment might be referenced as a fragment spread, but not used in code, but should cover the majority of the cases.

Reviewed By: leebyron

Differential Revision: D6229473

fbshipit-source-id: a8e8c48687414f074566b3aab25f8619cc4183b5