fix flatten transform for non-simple types
Summary:
When the fragment was spread in a place that is not a simple type, but rather a
non-null or list type, we didn't flatten the fragment. Fragments don't know
about these wrapper types, so this change is to just compare the raw types
instead of requiring the same wrapping types.

Reviewed By: leebyron

Differential Revision: D6417947

fbshipit-source-id: f371037775d5e7ea621efd59227b4534acb74fd6