Add CircleCI config for website deployment
Summary:
This is pretty hacky and copied together, but worked to deploy on my user account.

What probably needs fixing is configuration that only master is deployed as I saw circle run on the gh-pages branch.

We might also want to update the user name to some bot or something?
Closes https://github.com/facebook/relay/pull/2188

Reviewed By: leebyron

Differential Revision: D6421498

Pulled By: kassens

fbshipit-source-id: 4dac8f8e0a7fbcbe24730b04b3839ffdcd092a36