rename setRelayModernMutationConfigs to RelayDeclarativeMutationConfig
Summary:
These are not "ModernMutationConfigs", they're declarative configs used in classic and modern. Maybe someday we'll unify on an updater function in modern, but for now these are convenient in both worlds.

Next step is to also move the flow configs there so we don't have to import them from classic.

Reviewed By: leebyron

Differential Revision: D6608710

fbshipit-source-id: 518c4c67a6011f92a636d7555b447c3179795b00