[Async React + Relay] Fix issue where 'setState' fires too early
**what is the change?:**

There is logic in the contructor to fetch data, and sets 'onDataChange'
as the callback for when the data is ready.

With React rendering asynchronously, the 'onDataChange' could fire
before the component is mounted. This adds protection for that case,
adding a flag to track whether the component is mounted. Only in cases
that the 'onDataChange' fires before the component is mounted, we will
defer the 'setState' call until 'componentDidMount'.

**why make this change?:**

With React in synchronous mode, we can assume the component will always
be mounted by the time we call `setState` inside `onDataChange`, like
so:
```
```

Once React is rendering asynchronously, if for some reason it takes
longer than the data fetch, then the `onDataChange` will fire before the
component is mounted, and call `setState`. This would throw an error,
and is not supported. It is not defined how this would behave in the
future, but it would not be good.
```
```

**test plan:**
Ran existing tests and they passed.
I'd like to add a test which covers this specific situation, but that
will take a bit more effort and if this change doesn't hurt anything
we'd like to land it sooner.

**issue:**
internal task t25062790