Make graphql a peer dependency
Summary:
`graphql-js` doesn't work if there are multiple copies installed in a project, regardless of if they are the same version. Having graphql be direct dep makes it really hard to consume relay-compiler in additional tools, like GatsbyJs, because we need to depend on the package manager deduping any instances, which can fail it happen in plenty of instances.

CC KyleAMathews
Closes https://github.com/facebook/relay/pull/2271

Reviewed By: leebyron

Differential Revision: D6897139

Pulled By: jstejada

fbshipit-source-id: ff46834395ad88003f359e48467331fa983396d3