add sourceHash to query hash
Summary:
When a query is using `relay(mask: false)` and parts of the query are moved either into or out of the factored out fragment, we need to update the `sourceHash` in the generated query even if the flow type and AST of the query stays the same.

`sourceHash` is the formatted AST (minus comments) of the literal source code in the text. It's used to issue a warning if someone didn't compile the query.

Reviewed By: jstejada

Differential Revision: D7341988

fbshipit-source-id: c638658d954d400ec361faa8230eb174257846e8