Makes flow generated '%future added value' for enums optional
Summary:
As suggested by leebyron in #2351, this adds an option to the `relay-compiler`  so it doesn't add `%future added value` to each GraphQL generated enum. My reasoning is that you might have a strict control over your GraphQL service, so you'd have to handle any future value by rerunning the compiler, with the new version of the schema.

I'm not totally convinced by the name `strictFlowTypes`, but I don't really have much better now. My thought between this name is that the option would be a generic option that configures the generated flow types, and would not be specifically to enums.

Regarding test cases, I'm not sure of how I should tackle that, since the only tests are snapshots, and there are no tests with different configurations of the `RelayFlowGenerator`. I could however try to make that happen if you'd like!
Closes https://github.com/facebook/relay/pull/2368

Reviewed By: kassens

Differential Revision: D7343235

Pulled By: jstejada

fbshipit-source-id: a74bbea48708a1b432fdcc7267fbba5efd2b950e