Fix relay-runtime main entry for OSS (#2497)
Summary:
With gulp, we bundle the package that we'll end up with an `index.js` in
the root. This PR keeps that and updates the `package.json` to revert to
the old behavior.

Thanks AlicanC for brining up this error in https://github.com/facebook/relay/commit/9d71698d575b9e917cf5b3b0ad197858bc0af8dc
Pull Request resolved: https://github.com/facebook/relay/pull/2497

Reviewed By: jstejada

Differential Revision: D9076784

Pulled By: kassens

fbshipit-source-id: f29948c7e95a17e8eed15d793c3249f428c4b991