Add warning and do not load data in `refetch`, `loadMore`, `refetchConnection
Summary:
There may be a case with `RefetchContainer` and `PaginatioinContainer` when they may be unmounted, but intervals, timeouts, unresolved promises callback may call `refetch`,  `loadMore`  methods.

1. We need to warn that that's the case.
2. We do not need to load the data from the server

In this fix we are looking at the flag `isUnmounted`, to make sure we are not loading data when components are unmounted, we also show a warning message with an explanation.

Reviewed By: kassens

Differential Revision: D9019048

fbshipit-source-id: cbb74cd2128782b4c9ba77d6aa92b3a6cc659618