Use `holdGC` in the `RelayPublishQueue` to disable GC during optimistic updates
Summary:
Why do we need a GC hold?

There may be a situation when one of the optimistic updates deletes something from the store, for example, we call a method `RelayConnectionHandler.deleteNode(...)`  in the `optimisticUpdater`. Next, before receiving a response from the server garbage collection happens (for some reason, something gets unmounted...). GC remove deleted nodes. And then when we are trying to revert the optimistic update a deleted node is missing.

In order to prevent disappearing items from the store during optimistic updates, we need to disable GC if there any applied optimistic updates in the store.

Reviewed By: jstejada

Differential Revision: D9362884

fbshipit-source-id: bcc066cbd35a5699d674294a6e98b36191fcf714