Make network use GraphQLResponse for data; move executeWithEvents() out of core
Summary:
This removes the concept of `executeWithEvents()` and the related `StreamPayload` type from Relay core and into an internal-only helper module.

The idea of `executeWithEvents()`  is to allow a server to return non-GraphQL payloads that may be relevant to the application, but we have found that this is not typically used. The implementation of executeWithEvents() means that *all* network implementations are expected to return the non-standard StreamPayload type, but for convenience we then also have to accept plain GraphQL payloads and convert them (wrap the objects) to StreamPayload. Instead, we want to simplify the network abstraction to reduce complexity both for Relay core (just handle GraphQL payloads) and for network implementors (just return GraphQL payloads).

A quick summary of the changes:

* Remove Environment::executeWithEvents()
* Remove Network::executeWithEvents()
* Remove the StreamPayload and related types from Relay core
* Add all of these back in a new, internal-only helper module.

Reviewed By: ximyu

Differential Revision: D10439170

fbshipit-source-id: d8a445226f76391fc305673b7051fe019d9a9f1f