generate require calls for data driven deps
Summary:
This adds a small utility that allows us to mark a string as "code" by wrapping
it with a marker. This object can then be JSON.stringified per usual and then
post-processed to turn the specially marked strings into code.

We use this utility to generate require calls for data driven dependencies.

Reviewed By: jstejada

Differential Revision: D12933229

fbshipit-source-id: 004dda4d02351315760bb472c2b36388ef3b60f3