change @match syntax
Summary:
This change makes the parser unaware of the new match directive and instead moves it all into the MatchTransform.

A nice side-effect is that the IR type can be made a bit stricter always requiring the type.

I also noticed that passing the parent during the traversal isn't really helpful as it might be a conditional node that doesn't have a type and you'd have to look at some ancestor. The proper fix is to track the type when traversing down the tree. I'll leave this for a separate diff for now.

Reviewed By: josephsavona

Differential Revision: D13206085

fbshipit-source-id: 14e9d50cf84cc0505f61574ce95647aeedde10a6