Parse all ASTs -> IR in a batch
Summary:
We currently have a few places where IR nodes end up without a `type`. In some cases we're explicit about this and make the `type` nullable, in others we actually keep the Flow type as non-nullable and have fixmes. The main impact of this is that the IR is less reliable causing some defensive coding, but more importantly that we can't do some important checks early in the compiler pipeline. For example, `arguments` aren't currently checked (until persist time), but ideally they could be checked when the IR is constructed.

This diff changes "parsing" into a batch operation taking a full set of ASTs and transforming them in-bulk to IR. It doesn't yet do any additional cross-definition checking to address the issues with `type` fields, but it sets us up to be able to.

Reviewed By: kassens

Differential Revision: D13351443

fbshipit-source-id: 3d346bf887bea1501b8777ef5a4b68d056a5d0fd