Build variable definitions before AST=>IR conversion
Summary: This adds two related changes in GraphQLParser with the goal of having more complete knowledge over the types of the variables in scope in each IR definition. First, the AST=>IR transformation process is split into two stages: the parser now constructs a map of the defined variables for each definition before transforming that definition from AST to IR. This doesn't immediately allow recovery of additional type information but sets up a follow-up where values passed in `arguments` can be type-checked if the corresponding argument is defined on the referenced fragment (the idea being that `arguments` is processed in the second stage and can check the variable definitions computed in the first stage). Second, all places that expect only literal arguments (things like default values or Relay-specific directives that don't accept variables) are now parsed specifically as literals. This means the `_transformValue()` helper is now only used in places where the value's type is known, guaranteeing literals are checked to be valid.

Reviewed By: kassens

Differential Revision: D13515497

fbshipit-source-id: 2910570c7b4b7684396a730bc503db470e76ec41