support literals in @arguments
Summary: Adds support for passing literal values in `arguments`. In order to ensure that literals are valid we need to know their type, ideally when converting from AST => IR in order to give good, localized error messages prior to transforms. This is achieved by first building up a mapping of definition names to asts + variable definitions, and then converting all those asts + variable defs to IR. When a literal in `arguments` is encountered the parser checks that the corresponding argument is defined and if so parses the literal with the defined type. Noe that this is more strict than when a variable is used: this diff intentionally leaves variables in `arguments` alone in order to avoid thrashing generated artifacts. Validation for variables (and using the types) will be added in a follow-up.

Reviewed By: kassens

Differential Revision: D13515751

fbshipit-source-id: 99cbca94417efcfe4bcc9bfc46be160d5b716125