Synthesize refetch queries with `@refetchable`
Summary:
Adds initial support for `refetchable(queryName: String!)` on fragment definitions, which triggers the compiler to synthesize a refetch query for the source fragment. The idea is to reduce the boilerplate (and possibility of error) associated with manually creating a refetch query that defines the correct variables, spreads the source fragment, and sets all its variables to the correct values with `arguments`.

Notable features:
* Infers the correct set of variables to use for the query definition by scanning all fragments reachable from the `refetchable` fragment and finding all root variables that are transitively referenced.
* Synthesizes refetch queries for fragment on the query type, Viewer, Node, or concrete types implementing Node. Provides specific error messages if the schema does not satisfy the contract for each of these cases.
* Generates an `$id` argument for fragments on Node and validates that the `id` field is fetched.

At present the transform isn't integrated into the compiler. Follow-up diffs will integrate this into RelayIRTransforms and also update ReaderCodeGenerator to emit a link (lazy require?) from the fragment to the generated refetch query.

Reviewed By: kassens

Differential Revision: D13541268

fbshipit-source-id: a98385ae91b92b558242e7ad8e45cac575c56c1c