Refetchable ReaderFragment exports the refetch query on metadata
Summary: Modifies `ReaderFragment` to make `metadata` an exact object with the specific key/values that are used in practice, and then updates the code generator to rewrite refetchOperation names into calls to require the module itself. Also adds a `ReaderRefetchableFragment` that is a refinement of `ReaderFragment` where the refetchOperation is present.

Reviewed By: jstejada, kassens

Differential Revision: D13577240

fbshipit-source-id: 869b1da9ac4853dc499813f79b582e052aaf9af9