Use metadata to link generated Requests/SplitOperations with their source fragment
Summary:
We need a mechanism to link synthesized artifacts with the node from which they're derived so that the synthesized artifact can be generated into the correct location (adjacent to the artifact for the source fragment). Currently this can include `refetchable` refetch queries and SplitOperations from `match`but there are probably more to come.

This diff changes from using `SplitNaming` for this purpose - that relied on generating names like `<SourceName>$<annotation>`, which results in invalid GraphQL identifiers to instead put `derivedFrom` metadata on synthesized nodes, and use that to recover the original name when choosing where to write artifacts. SplitNaming is now renamed to just a utility for generated names for `match` normalization queries.

Reviewed By: kassens

Differential Revision: D13585975

fbshipit-source-id: d6eaad789cd86886701c0068bc107882c8f5d37e