Incremental 5/n: Initial normalizer support for Defer/Stream
Summary:
Add initial support for normalizing Defer/Stream nodes. Note that Defer/Stream incremental data delivery can be configured dynamically via the `if` argument. When disabled these nodes should be handled similarly to an InlineFragment by normalizing the child selections eagerly. In summary:

* When Defer is disabled, normalize its selections eagerly (no incremental chunks will arrive)
* When Defer is enabled, return metadata about how to handle response chunks (the path, label, and a selector w which to normalize)
* Always eagerly normalizes Stream nodes. If Stream is enabled, also returns metadata about how to handle subsequent response chunks.

Note that additional work is required to support the combination of `match` and defer/stream (or even nested defer/stream): in particular, correctly handling nesting will require passing in the parent path when normalizing a child chunk. I'll address this in a follow-up.

Reviewed By: jstejada

Differential Revision: D13795484

fbshipit-source-id: 536ae166045b5d2ad1557739f3c046994feb9de0