RM Owners 4/n: Update owner arguments for getSelector*() APIs
Summary:
In D13745761, I updated our getSelector*() APIs to optionally take a *single* fragment owner, in order to opt-in into consuming the variables from the fragment owner.

However, passing a single owner to `getSelectorsFromObject` or `getSelectorsFromList` was extremely cumbersome since it would require extracting a single owner from a set of fragment references and validating that they all had the same owner.

These changes to the API make it easier to use and pass singular or plural fragment references directly from the props that are naturally received by a fragment container.

Reviewed By: josephsavona

Differential Revision: D13873605

fbshipit-source-id: f27e38ef19592c5aa59f3393e1981d5ad5dfe8cb