@match/@module: eagerly process the match component
Summary:
Changes the way data-driven dependencies are processed. Before, it was inconvenient to distinguish between the cases where no selection matched vs when a selection did match but was still being loaded. This is now changed so that the `__module_component` field is eagerly processed when a `match` (or `module`) is reached during normalization. The result is that a fragment using these directives can resolve fully (no missing data, no suspending - unless due to some unrelated missing data) even  before the match data completes. When combined with `<MatchContainer>`, this means that the two cases will behave as follows:

* A selection matched: the parent component (using `match`) renders w/o waiting for the match field to process. It renders MatchContainer, which tries to render the selected component (and may suspend if it's not ready). When it is ready it renders the component, it may suspend if its data hasn't finished processing. Either way, we always just keep trying to render the matched component until it's ready, w/o ever using the MatchContainer `fallback`.
* No selection matched: the parent component renders MatchContainer, which notices that nothing matched and renders the `fallback` (which defaults to `null`).

Reviewed By: jstejada

Differential Revision: D14010824

fbshipit-source-id: 284a521f9e62e28618bc55b476642bd9e4b70615