QueryExecutor support for nested @stream fields
Summary: Adds support for nested `stream`. The only real runtime change is that instead of storing the root RecordSource so that the stream parent can be looked up, we instead store every record from every payload that appears as a parent of an `stream` field. The tests for the nested case all start from the state at which a parent has already resolved. The main case that might seem not to be tested is if a nested payload shows up before its parent, but that's actually just equivalent to the more general case (already tested) that we throw a reasonable error in the case of getting a payload we don't know how to handle.

Reviewed By: jstejada

Differential Revision: D14096113

fbshipit-source-id: 8de619e7322503404b4e1cf537f753bc126eaa8d