Initial @stream/@connection integration
Summary:
This implements initial support for streaming connections (concretely `connection` where the edges field uses `stream`). Note that connections are implemented by effectively creating a client-only field in the schema and populating this field with merged results of the initial/pagination queries on the source server field. Prior to this diff, `stream` payloads would update the server connection fields but nothing would trigger an update of the derived client field, which would therefore always appear to be empty. With this diff, QueryExecutor now stores the relevant HandleFieldPayloads from a parent payload and replays them when processing each (streamed) child payload. This ensures that derived fields correctly see each new edge.

Note that the connection compiler transform currently adds some fields automatically, such as pageInfo, which we would prefer to defer when the edges are streamed. This causes "Payload did not contain a value for field `pageInfo: PageInfo`" warnings at present (silenced in the unit tests here). Follow-up diffs will add tests for more edge cases (and fixes as necessary) as well as adding options to the `connection` directive to support streaming with the pageInfo field deferred.

Reviewed By: jstejada

Differential Revision: D14143498

fbshipit-source-id: c6338836c6a9c7b2e4ec84f456a50316b2b8c15e