Support using field handlers on the same field as @stream
Summary: Updates QueryExecutor to support handle fields directly on a `stream` field. Prior to this change the handler would not get called when new items arrived at the source field, and the derived field would only show the items that were part of the parent payload. With this change, QueryExecutor now stores HandleFieldPayloads on the stream field itself and replays them each time a new edge arrives.

Reviewed By: jstejada

Differential Revision: D14226862

fbshipit-source-id: 839dc4da37725bd479f8da917577fdcfc466f073