Ownership: Fix selector comparison when using fragment owners
Summary:
When fragment ownership was enabled under the PREFER_FRAGMENT_OWNER_OVER_CONTEXT feature flags, I missed an important update to `areEqualSelectors` to check if the owner in the selectors is different.

Currently, areEqualSelectors is used to determine if a the selector used to read a fragment inside RelayModernFragmentSpecResolver should be updated when setting new variables or new fragment refs. With fragment ownership enabled, we would return true from `areEqualSelectors` even if the owner on the selectors was different, which means that we wouldn't update the selector and would end up propagating stale variables down the tree via a stale owner. This was evidenced by the issue described in T40967852, in which rendering a tree with new fragment refs wasn't properly updating it since data was read with stale variables.

Similarly to what we did in D14206413 to include the owner identity in the fragment identity for our experimental apis, areEqualSelectors needs to return false if the fragment owner has changed.

Reviewed By: josephsavona

Differential Revision: D14239824

fbshipit-source-id: f8966141cfb9dee58628bdd425b6449201273422