Add PublishQueue interface
Summary:
This diff is adding an interface for Publish Queue, and ability to pass an instance of PublishQueue directly to the Environment.

Why do we need this? We need to keep track of all incoming operations that may affect fragment owners to mark them as pending.

For that purpose, we could wrap Publish Queue with a dedicated module that will keep track of all incoming operations (but that's TBD) and how they are affecting fragment owners.

Reviewed By: josephsavona

Differential Revision: D14325114

fbshipit-source-id: f274525d0a1cd7426411dc6dbc56f09c9a70ea6a