Feature flag for incremental data delivery
Summary:
Adds a RelayFeatureFlags.ENABLE_INCREMENTAL_DELIVERY that toggles the feature, default to false (disabled). When disabled the following checks occur:

* Relay compiler *requires* defer/stream to have an `if: $RELAY_INCREMENTAL_DELIVERY` argument when the above flag is false.
* QueryExecutor throws if defer/stream are used where the `if` condition is set to true (ie, the directives themselves are allowed, they just have to be disabled if the feature itself is disabled).
* The FB fetch function errors the request if the query variables include `$RELAY_INCREMENTAL_DELIVERY` and the variable is true (when the feature is enabled, the variable can be set to either true or false).

The idea is that relay compiler will use the "off" mode throughout the rollout period (requiring the specific `if` variable until we want to enable this more broadly), since this restricted form is just a subset of what's supported at runtime. At runtime, we can still enable the feature for a subset of users.

Reviewed By: kassens

Differential Revision: D14358766

fbshipit-source-id: a89fcb3238bad2dc53b4f9f3cf26ea7c013892d9