Trick React into thinking a container is a class
Summary:
The factory component pattern is about to be deprecate in:

https://github.com/facebook/react/pull/15145

This change makes it so that React will call this constructor with `new` as if
it is a class component instead of just calling it as a function component.

There might be a slight perf penalty to doing this my understanding is that this
is not the preferred path anyway.

Reviewed By: kassens

Differential Revision: D14516681

fbshipit-source-id: 751890a98639687b208633f2e4c8e71e5df60776