Add experimental Chrome Console console inspector
Summary:
With this change, each Relay Environment get a new method in dev mode only:

```
// defaults to the root object
myEnvironment.DEBUG_inspect()
// or
myEnvironment.DEBUG_inspect('some-id')
```

that returns a Proxy object that allows to explore the store as if it was denormalized. This works best if "Enable custom formatters" is enabled in the Chrome Developer Tools settings.

This is pretty experimental and we'll see if it proves useful.

Reviewed By: alunyov

Differential Revision: D14553843

fbshipit-source-id: cca104f587bf7a3d8cb0577404301e7e01dd5fe3