Unused argDefs causes runtime error 2/2: Fix
Summary:
Fixes an issue that occurs when a fragment references a global variable inside arguments and the fragment it is passing the argument to doesn't end up using the corresponding argument: in this case, the compiler doesn't see the global variable ever being used, and doesn't check that there is a corresponding query argument definition. Without a query argument definition, Relay runtime previously did not populate a default value for the variable, which would then throw in RelayStoreUtils.getStableVariableValue due to the variable name not being defined as a key in the variables object. The fix is to explicitly set the value of these global variables to `undefined` so that accessing the value still produces the same result, but it passes the hasOwnProperty check.

A subsequent diff will add stricter checking of args/argDefs (that all args are defined, and that all argDefs are used).

Reviewed By: jstejada

Differential Revision: D14729160

fbshipit-source-id: a0a45680094d5b71a4b0e2840a83b62a0e0f60c6