QueryExecutor optionally schedules work
Summary:
Allows QueryExecutor to optionally schedule processing of payloads: "root" payloads, incremental payloads, and data-driven dependency follow-ups. The use-case is to allow us to experiment with rendering content earlier. Today for example if two streamed items are already ready (prefetched) when a QueryRenderer mounts, Relay will process both items and then React rendering will occur where the list has two items. The scheduler would allow controlling processing so that even if multiple items are ready, Relay yields to React and gives it a chance to render e.g. the first item, then let Relay process the next item, then yield again and render the second item.

Note that this is designed such that the *order* in which payloads are processed is the same as before, though we may experiment with prioritizing some payloads before others in the future.

Reviewed By: jstejada

Differential Revision: D14895483

fbshipit-source-id: c964fd3b96fc938fe304550436824add510043f7