Client Extensions 2/n: Add ClientExtensionsTransform (not enabled yet)
Summary:
This commit adds a new ClientExtensions transform to our compiler transforms.

This transform will do the following:
- Group all sibling client-only selections (defined on a client type, or defined as an extension on an existing type) under a single ClientExtension node.
- Ensure that there are no nested ClientExtension nodes; anything selected under a ClientExtension will automatically be treated as a client-only selection, so we don't need to create extra ClientExtension nodes

The ideas is that these measures will minimize the amount the amount of extra nodes we need to generate in our generated artifacts.

Additionally, this commit updates GraphQL IR printer to print out ClientExtension nodes so that we can run the tests and generate snapshots for the new node type; however, when printing, we will actually strip out these selections.

The next diffs in this stack will enable the transform and update our runtime traversals to handle them

Reviewed By: josephsavona

Differential Revision: D14868479

fbshipit-source-id: c83ce018eb9d12fb41194b73e7cae6c61d5cf897