Client Extensions 3/n: Add SkipClientExtensionsTransform (not enabled yet)
Summary:
This commit adds a new SkipClientExtensionsTransform transform to our compiler transforms. This transform will remove any ClientExtension nodes it encounters, as well as Fragment definitions on client types, and /depends/ on ClientExtensionsTransform to have been run first.

This transform matches the behavior of the existing `SkipClientFieldsTransform` (see tests), and the new transform will eventually replace the original further up in this stack, when the new ClientExtension transforms have been fully enabled.

Reviewed By: josephsavona

Differential Revision: D14931034

fbshipit-source-id: d2c3ea9105b106436d226cf54871b90c3668ab8b