Split up RelayModernEnvironment-test
Summary: This is a naive split of RelayModernEnvironment-test into one file per top-level  `describe`. The top of the file - imports, locals, beforeEach, afterEach - is copied identically to each test and then minimally modified to satisfy lint and reduce duplication (e.g. merging multiple beforeEach blocks into one), in order to make it easier to verify that this is all just copy-paste. This makes the environment tests faster since they can run in parallel and easier to edit (instead of one gigantic file).

Reviewed By: alunyov

Differential Revision: D14943762

fbshipit-source-id: 34331472e0d84783ba36d7c69809d64ecc8c253b