Community Resources: add link to Relay examples repository (#2707)
Summary:
Hello guys! 👋

We had internal Relay workshop some time ago and we decided to record some parts and share it with the community along with short documentation as well as with working code. I believe this could serve as a good source of information even for other people working with Relay.

Thanks for considering it.
Pull Request resolved: https://github.com/facebook/relay/pull/2707

Differential Revision: D15183260

Pulled By: jstejada

fbshipit-source-id: a303ff23ad5152e77bc032a88c465f7c16037a6f