Ensure SplitOperation artifacts are deterministic w namespaced 3D results
Summary:
Fixes a bug that I introduced in D15349734 (namespace 3D results by parent document name) that caused compiler output to be non-deterministic when multiple documented fetched the same fragment using `module`. The SplitOperation (generated in the `...$normalization.graphql.js` files) contained `js` field selections for the fragment/component, and these would end up with an alias based on the most recent parent document from which the SplitOperation was derived.

The fix is to no longer generate normalization nodes for these js fields in SplitOperations at all. Instead, ResponseNormalizer is responsible for normalizing both the component and operation fields. This also means we can save bytes: two less fields to represent in the generated ASTs.

Reviewed By: jstejada

Differential Revision: D15390037

fbshipit-source-id: a9c68d7318c8852d89d3cf64b6a218604667b89a