Support additional arguments on @match fields
Summary: Changes RelayMatchTransform to support using at-match on fields that have arguments present. Previously any existing arguments were dropped and we inconsistently errored if the field *definition* included other arguments besides 'supported' due to a bug. Now, the compiler validates that there is a defined 'supported' argument on the field definition and that that argument is not explicitly provided, but it preserves other args.

Reviewed By: alunyov

Differential Revision: D15565607

fbshipit-source-id: e549f1c4d7721fe491b0b29ac825a454d0f50f15