Hack to forward old urls to relay.dev
I tried adding a forward using GitHubs domain setting, but that clashed
with some other config and triggered an infinite redirect.

This is just a terrible hack in a terrible place to add a forward to
most pages (all that use the footer, this excludes the GraphQL spec
extension pages).