Support for dynamic connection keys (behind feature flag)
Summary: Implements support for dynamic connection keys. Relay currently supports the `key` argument to namespace the results of a connection field - but because this value is static, if multiple instances of the same component (and therefore fragment) are subscribed, they will share the underlying connection state in the store. There are some use-cases where applications would prefer to have data namespaced on a per-instance basis. We plan to investigate this in a follow-up, but as short-term solution this diff implements support for dynamic connection keys. In addition to the still-required static key, a fragment may specify `dynamicKey_UNSTABLE: $someVariable` (must be a variable) whose runtime value is used to compute the storage key (in addition to the filters and static key). This is behind a feature flag as we intend to investigate alternatives in the near-term.

Reviewed By: jstejada

Differential Revision: D15704185

fbshipit-source-id: 3e8791d3ad0f28cb5334e241ab924867225847a4