Add readInlineData for @inline
Summary:
A while back we added a `relay(mask: false)` hack on fragment spreads that inlines the fragments verbatim. This helped for product code that wanted to use non-React functions called from React. Unfortunately, that approach breaks data masking and also doesn't work with `arguments`.

We could have added a `readFragment(environment, fragment, fragmentPointer)` API that could be used from functions. The problem with that is that it doesn't always work as expected when called from  a React render: The Relay container only subscribes to the data it read directly, if the function accesses a deeper object it wouldn't be called again when only that deeper object changes since the component isn't subscribed to it.

We could've also added an API like `subscribeFragment(environment, fragment, fragmentPointer, callback): Disposable`, that is called whenever the data changes. This seems pretty unergonomic for the use cases I've seen though and puts a burden of correctly disposing on the API consumer.

Finally, this proposes a new directive `inline` on fragment definitions. Similar to `relay(mask:false)` a fragment with this directive is read together with the parent fragment (and therefore subscriptions are setup with the parent). But unlike unmasking, this hides the data in a new kind of fragment pointer similar to other fragment spreads. This fragment pointer can be read with a new util function wrapper.

Reviewed By: alunyov

Differential Revision: D7304389

fbshipit-source-id: 86bbbf59718887824d524c9a50765ae69a3802c9