New Connection Model 1/n: Add ConnectionField IR kind
Summary:
## Context

Over the years we've iterated on various approaches to handling connections and learned a lot about what works and what doesn't.

* Relay Classic deeply integrated connections into the implementation, but the approach wasn't flexible (users couldn't really override how various lifecycle events - pagination, refetching, etc - were handled. Further, there was a single  instance of each connection, leading UIs to effectively share connection state by default.
* With Relay Modern we were reluctant to integrate connections deeply into the core, and instead came up with a split approach in which an `connection` compiler  transform and runtime ConnectionHandler were implemented in terms of a more generic "FieldHandler" abstraction. This enabled much more flexibility than classic, and via the "key" and "filters" properties (and recently the dynamicKey_UNSTABLE property) allowed having multiple distinct views of the same connection. Overall this solution ultimately allows products to achieve most use-cases, but is  a) more challenging to use than we'd like for simple cases (too many options to configure) and b) not quite flexible enough for complex cases (not quite enough integration points even if you do implement a custom handler and use the various filter/dynamicKey options).

We've come to appreciate that **connections are a fundamental primitive in GraphQL.** Most programming languages have some core type that represents "a sequence of values that can be iterated over" (sometimes called Iterable, Sequence, etc) as well as a type representing a particular instance of an iteration (often called Iterator) - in addition to being part of the standard library, these types often receive special language support (e.g. loop syntax that desugars into usage of an Iterator). Databases similarly offer cursors over result sets.

GraphQL clients requesting a subset of a data graph from a GraphQL server also need to be able to request and iterate over sequences of values, *without* loading the entire (possibly effectively infinite) list up-front. This is fundamentally asynchronous and distributed in nature: the list could change on the server,  much like an Array being iterated over can (in many languages) be modifed during iteration. This means the API must include different information from a typical `Iterator` to help resume iteration without a persistent client/server connection (or a stateful server more generally). This inevitably leads to roughly the shape of the Relay Connections Specification: some way to get a page of results w metadata about that page's place in the overall sequence, and then later fetch the "next" (or previous) page or refetch a previously fetched page to get fresher data, using the page metadata to reconstruct a merged view of the items on the client.

Our takeaway from this is that we should support *some* variation of a connection-like abstraction as a first-class concept in Relay. Just as different languages have slightly different `Iterator` APIs, we expect different schemas to have slightly different representations of connections. Ultimately, we'd like our implementation to be flexible enough to support more than just the exact format of the Relay Connection Spec. So we're designing with that as a long-term goal in mind, but initially supporting exactly that spec. *Future diffs will lay out the exact approach in more detail.*

Note that for the foreseeable future this will be *in addition to supporting the current connection approach*. As we implement the new model, we'll figure out a migration path for existing code.

## This Diff

As a first step, this diff adds a new IR kind `ConnectionField` representing connections which includes the same information as LinkedField plus the name of the `resolver` *module* to use. For now we only add the IR kind and handle it in transforms - the next diff will add a transform that actually creates instances of this new kind.

In theory we could use metadata to store the extra information relative to a plain LinkedField, but adding a new IR kind helps ensure that we're handling the field appropriately in various transforms.

Reviewed By: kassens

Differential Revision: D15891507

fbshipit-source-id: 2637d11ade50d92132b53347e9bb6c4f134a2936