merge RelayCodeGenerator-test into compileRelayArtifacts-test
Summary:
The `RelayCodeGenerator-test` is a test that's an integration test of compilation steps and codegen, but only applying few transforms. This results in IR nodes arriving in the code printer that we don't expect there.

We already have a proper compiler integration test that includes codegen (`compileRelayArtifacts-test`) so here I'm just merging the fixtures that didn't exist in the full integration test over and delete `RelayCodeGenerator-test`.

Reviewed By: josephsavona

Differential Revision: D16364111

fbshipit-source-id: 73e9129dd552f40de6b2d24f2d0c87c9f36b1d7a