Convert OwnedReaderSelector to union type (first-class plurals)
Summary:
Changes `OwnedReaderSelector` to be a union of single/plural variants, and updates the various relevant helpers accordingly. As part of this, I realized RelayModernSelector-test.js wasn't Flow-ified yet and did so, which helped me fix all the failures of missing 'kind' values :-)

Note that this introduces an extra indirection for plural selectors: given that they are used much less often this shouldn't be a problem in practice.

Reviewed By: kassens

Differential Revision: D16471511

fbshipit-source-id: d803c162e413eb7e67978ebaf179d59d793c6983