RelayParser: validate defined fragments
Summary:
Now that we don't have any references to fragments defined in classic mode (which were invisible to the modern compiler) we can enforce that all fragment spreads are for defined fragments. We've already done this in some places, this adds the check to RelayParser to find these issues early in compilation.

Note: I wanted to use the didYouMean and suggestionList helpers from graphql here to give suggestions in case of typos in the fragment name, but it looks like they aren't exported. We can follow up to somehow get these utils (export them in graphql-js or  add local alternatives) in a follow-up, there are other places where we could use them such as undefined field/fragment arguments.

Reviewed By: kassens

Differential Revision: D16527841

fbshipit-source-id: 6cffde3a5426ac1ee0bcb05621a731569c6e76a8