More context for "variable not in scope" errors
Summary: In general the compiler infers and validates the set of root variables that are expected to be in scope for a particular fragment. However, using `relay(mask:false)` bypasses this which means that unexpected global variables can be referenced at the point ApplyFragmentArgumentTransform runs,  before we do a follow-up pass to  verify globals in RefineOperationVariablesTransform. Since we want to deprecrate masking anyway, rather than a full fix (again, we already do a final validation pass anyway) this diff  provides more context about which query was missing the variable.

Reviewed By: kassens

Differential Revision: D16627852

fbshipit-source-id: efa8c840123ac286a1b4be3180fdf023f5d2619d