Run update-spec in website package to update connections spec (#2819)
Summary:
As referenced in https://github.com/facebook/relay/issues/2457, the connections spec markdown file is up-to-date but the published artifact related to it is not. This PR just runs `yarn update-spec` in the website package in order to get the spec site back in sync with the current version of the spec.

This was the simplest solution to the immediate problem, but is not long term. I could see a more robust solution that removes `website/static/graphql` from the git repo entirely and moves `yarn update-spec` (maybe renamed to `yarn generate-spec` or something) into the circleCI config so that the html artifacts are generated during the publish and thus cannot get out of sync, but I did not want to overstep (though could easily wire that up as a follow-up if desired).
Pull Request resolved: https://github.com/facebook/relay/pull/2819

Reviewed By: jstejada

Differential Revision: D16705673

Pulled By: kassens

fbshipit-source-id: bc79b5dda7682c0259e044d89b1fed995ff12327