Simplify optimistic updates, move optimistic normalization to QueryExecutor
Summary:
We currently normalize payloads in a few places: server payloads and optimistic responses *from the network* are normalized in QueryExecutor (which handles any 3D follow-ups), but optimistic mutation payloads were normalized in PublishQueue. This diff refactors the structure of optimistic updates so that PublishQueue only deals with already normalized data (e.g. a RecordSource + FieldPayloads), and ensures that all optimistic responses are normalized in QueryExecutor.

A follow-up will build on this to handle 3D payloads in optimistic responses.

Reviewed By: kassens

Differential Revision: D16680058

fbshipit-source-id: d1da3426e2b933f2de4bf01db39b850adc757c35