Environment: make PublishQueue not configurable
Summary: Implementing support for connection resolvers with optimistic update/rollback will likely involve changes to the responsibilities of PublishQueue and Store. We made PublishQueue injectable in RelayModernEnvironment but a) we don't rely on this internally and b) it limits our ability to iterate on the implementation. Since we don't use this and it was only recently added, I'm removing it here to free up the ability to remove the concept of a PublishQueue that is separate from the environment and store.

Reviewed By: alunyov

Differential Revision: D16796416

fbshipit-source-id: 7dbcb92f054ef0c7714aaeb6e685f6436361e3f7