New connection model 5/n: 'update' event for changes on edge data
Summary:
The first pass of the new connection model handled new events and data updates by always replaying connection events in full, rebuilding the connection state from scratch on every change. In addition to the performance limitations of that approach, it also means that connection resolvers couldn't detect changes to distinct edges, since this would appear as just replaying old events w different data. Ideally resolvers would be able to see what specific edges changes and base decisions on this.

The new approach implemented here does the following:

* On Store::run(), instead of replaying all events for each connection subscription we now update the subscription with just the new events, starting from their previous state.
* But before iterating the new events, *if* any data for the connection has changed, we first process an "update" event that contains only the changed edge data. This allows connection resolvers to see exactly which edges changed.

The side effect of this is that connection resolvers will generally need to maintain state containing an ordered set of edges (or pages) by edge ID, a mapping of edge ids to data, and a projected set of edges (though this could be avoided by mapping edge ids to edge data in the render function).

Reviewed By: jstejada, kassens

Differential Revision: D16776455

fbshipit-source-id: e5d59d4273f3964e13cabbf016b604076e4d0b77