FlattenTransform: remove hot object spread
Summary:
This code is extremely hot. The object spread showed up in a line-by-line profiler using the Node remote profiler. This makes sense as an object spread needs to iterate properties whereas the inlined version directly knows the hidden class to create.

I was thinking if there was some more general solution we could use, but nothing convinced me so far:
1. Create functions like `createLinkedField(field1, field2, ...)` this could be confusing with many arguments, but guarantee same hidden class and no spread. Painful to update.
2. Add a babel transform to inline `updateLinkedField(oldLinkedField, {name: 'new'})` into what I did here. Not sure I'd like to add more custom transforms into our mix :/

Reviewed By: josephsavona

Differential Revision: D17046291

fbshipit-source-id: 935de03f3bfc5a3b8bc68b886b3e3661d31aa07f