New connections: support adding edges in mutations incl. final/optimistic w rebase/revert
Summary:
Implements support for inserting individual connection edges via final/optimistic updater functions and full support for rebasing/reverting of optimistic connection updates. The basic idea is that in PublishQueue::run(), after all "final" (non-optimistic) updates have been applied it takes a snapshot of all the connection subscriptions in the store,  then runs optimistic updates if present. When an optimistic update has to be rebased, PublishQueue restores the connection state (along w publishing its backup RecordSource). RelayModernStore now eagerly updates connections on publish/publishConnectionEvents, which sets a `stale` flag on the subscription if data changes. Then in `notify()` the store simply iterates connections and calls the callback for any subscription that is stale.

My main concern with this implementation is the size of the connection backup object - it's currently O(total connections),  not O(changed connections), unlike the backup RecordSource which only contains backups for records that changed. It's definitely possible to reduce the size to O(changed), but I'll do that in a follow-up (one idea: publish() and publishConnectionEvents() could return a list of connections that changed. Another option is to track both the committed and optimistic results on each subscription object).

Planned for follow-ups:
* O(changed) snapshots
* GC integration
* DataChecker integration
* Public API (hook)

Reviewed By: kassens

Differential Revision: D16956109

fbshipit-source-id: 3b6ccc527d71ab3972e09068ddb738016045dab9