Remove unused RecordSource::load()
Summary: Removes the unused `load()` method on RecordSource. The original intent of this was to support having RecordSources that could collect (during GC) records and purge them to disk, and then reload them back into memory if later needed. We haven't ended up doing that so let's remove the method for now.

Reviewed By: kassens

Differential Revision: D17207314

fbshipit-source-id: 754bcb1eccf1596564e5424964c6d5075c7f03ba