New connections: optimize connection snapshot/restore
Summary: Optimizes the process of snapshotting and restoring connections: instead of creating a full copy of the events and subscription-state maps, each entry now stores optional backup information. This still requires an O(n) traversal of the event and subscription-state maps, but that's pretty similar to what we already do for regular subscriptions (iterating all subscriptions to find those that are invalidated).

Reviewed By: jstejada, kassens

Differential Revision: D17233310

fbshipit-source-id: 23633948a531c441c0e33e61004c258a9f75106d