New connections: GC/DataChecker integration
Summary:
Implements GC and store-rendering support for new connections by supporting the `Connection` node type in ReferenceMarker/DataChecker. As an optimization relative to `connection`, ReferenceMarker tracks which connection ids it has already visited and avoids traversing the same connection multiple times. This is possible because the connection id derives from the `label`, which the compiler ensures is a unique identifier for a given connection field within the application's GraphQL selections. So it's still possible that the same edges (ie "feedback X's comments") will be traversed multiple times *if selected by two entirely different components*, but notably edges will not be re-traversed for the same connection instance that happens to be reachable by multiple queries (as often happens when a connection is paginated, since it's reachable from the initial query and all follow-ups). This also correctly handles edges added via mutations, since the set of ids to traverse is driven by the edge ids that appear within a connection event.

## Implementation Notes

* Store now exposes a way to get the connection events for a given connection ID (getConnectionEvents_UNSTABLE)
* ReferenceMarker/DataChecker use this method to find all the connection events for each connection, then traverse all the edges that have been fetched on that connection.

Reviewed By: jstejada

Differential Revision: D17404458

fbshipit-source-id: ea56ccd9cecdccd7b0f0c4334d5f644bc565a08b