Refactor Connection IR/AST representation, support metadata fields and improve generated Flow types
Summary:
The motivating change here is to support reading metadata fields on connections such as `count`. This was tricky with the previous structure of ConnectionField IR/AST nodes, since the metadata fields would fall underneath the connection node, which had special logic. What we really want is to treat the metadata fields as regular fields, and just isolate the edges/pageInfo since they have special meaning.

This diff accomplishes that by refactoring the IR/Node structure:

```
comments(first: 10) connection_resolver(...) {
  count
  edges { /* edge fields */ }
  pageInfo { /* pageInfo fields */ }
}

// previously turned into this Node structure:
ConnectionField {
  name: "comments",
  args: [...],
  selections: [
    ScalarField { name: "count" }

  ],
  resolver: "...",
  label: "...",
}

// now it turns into:
LinkedField {
  name: "comments",
  args: null,
  selections: [
    ScalarField { name: "count" }
    Connection {
      resolver: "...",
      label: "...",
      edge: LinkedField { name: "edges", ... }
      pageInfo: LinkedField { name: "pageInfo", ... }
    }
  ],
}
```

Note that in the runtime ASTs `comments` is a regular LinkedField again, and `count` is a scalar within it - this means that everything around edges/pageInfo is written and read normally, and only those two fields are special-cased. In the IR the Connection type has a regular `selections: Array<Selection>` property, but in the concrete nodes it has edge/pageInfo fields only, which facilitate reading out per-edge information (we previously scanned the selections to find the edge field at runtime, now that happens at compile time).

Also note that the LinkedField has its arguments stripped, since the only argument was `first` and that doesn't affect the identity of the connection at all. The argument stripping is a bit tricky - we only want to strip connection args in the normalization/reader ASTs, but not in the query sent to the server - so this diff keeps `ConnectionField` around but only in the IR. In codegen ConnectionField is translated as a LiinkedField and connection args are filtered out.

This change makes it easier to produce accurate flow types (so simple i just did it here). It will also make it easier to implement GC and DataChecker integration: again, the connection and metadata fields are checked normally and only edges/pageInfo will have to be special-cased.

Reviewed By: jstejada

Differential Revision: D17356747

fbshipit-source-id: 0fd11fa78b943ccc2a167627335c92688a2adab7