New connections: Flow types include edge shape
Summary: Improves Flow-type generation for `connection_resolver`, outputting the shape of the edge selections as a type parameter to `ConnectionReference`. The generated Flow types are still not fully valid - currently ConnectionReference takes a second type parameter that can't easily be provided by Relay Compiler, the ConnectionResolver state - but a follow-up will address this (see D17435523 which removes the second type param to ConnectionReference, making the types generated here correct).

Reviewed By: alunyov

Differential Revision: D17433752

fbshipit-source-id: 33cd9f454162b50551a39252c767b67a25d07380