New connections: provide resolver at runtime
Summary:
Changes from embedding the ConnectionResolver in the query statically (which requires passing it through the IR and into runtime ASTs) to instead requiring that it be passed at runtime to lookupConnection()/subscribeConnection(). Originally I thought we might want to do more resolver-specific customization in places like DataChecker or GC, but on further exploration I realized we can do something reasonable by default (see previous diff implementing Checker/GC integration), and we can always consider adding the ability to statically provide the resolver back later if we need it.

For now, passing the resolver at runtime simplifies things. In particular, it means that our generated Flow types only describe the data that's in the store. It's the call to lookup where the user is responsible for passing a resolver that is compatible with the data. For example, it would be an error to pass a resolver that needed access to edge `cursor`s and to forget to select that field.

Reviewed By: jstejada

Differential Revision: D17435523

fbshipit-source-id: 89e1e2e3c7ce3f681f3d9ea59b926935994dae90