New connections: make label required, enforced globally unique
Summary: Makes the 'label' argument to `connection_resolver` required and enforces that the user-provided value is globally unique (equal to the name of the document where it is defined or prefixed with "<DocName>$"). The idea is that, similarly to keys on existing `connection`, various APIs may require developers to specify which connection they want to target, and it's difficult to do so if the internal connection label doesn't match what the user provided.

Reviewed By: kassens

Differential Revision: D17506699

fbshipit-source-id: fdca4c1760f437d0a78d2f63afea05ff5755b3f9