Fix disposing of cached query resource when unsubcribing from network request
Summary:
In D17275072, we fixed an issue with the following scenario:

1. render and suspend on query A
2. switch to query B while query A is suspended, this will cancel the network request for query A
3. switch back to rendering query A

In step 3, query A would remain indefinitely suspended because when we canceled the initial request for query A, we didn't clear the cache entry for query A in QueryResource.

The fix in D17275072 was to unconditionally delete the cache entry when the network request was cancelled, however that fix is incorrect because it breaks the case when 2 different components are rendering the same query: if the first component cancels the request or is unmounted, it would delete the cache entry even thought it was still retained by the other component.

Additionally, this was only a problem if it happened /before/ the component had committed (i.e. while it was suspended), since after it commits, it will correctly cancel the request and delete the cache entry when it's disposed.

Given that, the fix in this commit is instead to dispose of the temporary retention if the request is canceled before the query resource has been permanently retained.

Reviewed By: josephsavona

Differential Revision: D17484308

fbshipit-source-id: 5db819c28ae6b0e63c67cc955a098c77078811bc