Calling getPromiseForRequestInFlight should not prevent request from being canceled
Summary:
Currently, fetchQueryInternal will dedupe identical requests in flight and use a ReplaySubject to replay all events that have occurred to new subscribers that attempt to start a network request that is already in flight.

The way it currently works, in order for the underlying request to actually be canceled at the network layer, all of the subscribers need to unsubscribe from the request.

This is usually what we want, /except/ for when we're calling one of the helpers just to check if the request is in flught, or to get a promise for a request in flight. However, as of right now, if we call `getPromiseForRequestInFlight` and the original caller for the network request tries to cancel the request, the request wont actually be ever cancelled since we attached another subscriber to create the promise for the request in flight.

Instead, the expected behavior should be:

1. If the original request is canceled, the underlying network request should actually be canceled, even if we called `getPromiseForRequestInFlight` was called.
2. The promise (or observable) for the request in flight should resolve when the underlying request is canceled (which is not currently the case either).

This commit fixes 1. by using a separate subject to broadcast events from the original request; open to suggestions for a better alternative. 2. will be fixed separately.

Reviewed By: josephsavona

Differential Revision: D17512468

fbshipit-source-id: 37f95d8db41f3d53a9487a6a6eb664bd415d5961