remove old RelayNetworkLogger
Summary:
This logger is being replaced with the event based logging on the Environment
which will allow us to have richer and more contextual events logged with more
detail on what's going on. Wrapping the actual network request doesn't work well
as some "requests" might never actually end up on the network when they can be
fulfilled completely from the store.

Reviewed By: josephsavona, alunyov

Differential Revision: D17602025

fbshipit-source-id: 84f5c7d3aa9d1e4baf60fb4abe28be7435812ddd