Refactor all graphql-js usages to use Relay Schema Interface
Summary:
This diff is pretty straightforward.

The first major change is in `GraphQLIR.js` where all GraphQLTypes replaced with the opaque TypeID from the Schema module.

This change enforces, triggers a waterfall of updates (due to Flow errors) in Relay Transforms, Code Generators and RelayParsers (and other related helpers/modules).

Next change is in the `GraphQLCompilerContext` which no longer have references to the `server` and `client` schemas. Instead, it receives an instance of Schema, and expose it via `getSchema()` method.

To correctly handle this change, I have to update unit-tests in the compiler, so they are getting flow coverage.

Reviewed By: josephsavona

Differential Revision: D17433238

fbshipit-source-id: 582618875a6612f30deb51512c0421b9b373abfd