useFragmentNode: copy impl of useFragNodes to useFragNode
Summary:
This diff is part of a stack to remove some layers of extra work from
`useFragmentNode` which is used for the `useFragment` hook. Before this stack
`useFragmentNode` called `useFragmentNodes` with a single key object which
leads to additional computation of a `fragmentSpecIdentifier` which isn't super
cheap as it uses `JSON.stringify` under the hood on the `fragmentIdentifier`
which also uses `JSON.stringify`. After this we'll at least only stringify once.

 ---

This step just inlines the implementation of `useFragmentNodes` with all the complexity.
The following diffs will remove the extra work that's no longer needed for easier
understanding of the changes.

Reviewed By: jstejada

Differential Revision: D17763840

fbshipit-source-id: 16017f0a413a8744ebdaa71566e39404771a6959