Include LICENSE in every package (#2888)
Summary:
The `once()` call seems to memoize that it was already copied, but this memoization caused the license to only be included in the first package we're building.

I don't think it hurts to copy the `LICENSE` on every build, so just removing the `once()` call here.
Pull Request resolved: https://github.com/facebook/relay/pull/2888

Reviewed By: tyao1

Differential Revision: D17982136

Pulled By: kassens

fbshipit-source-id: 94d55f62c8dfb9b26e39f4619c87ac496d1d8c8a