Publish master releases from CI to NPM (#2889)
Summary:
Adds a GitHub action that runs only on pushes to the master branch on facebook/relay. This step pushes a build to the master tag on npm which can be installed with `yarn add relay-runtime@master` or similar.
The master releases have a version like `0.0.0-master-62d0ae50` which includes the Git commit hash for easy reference.
Pull Request resolved: https://github.com/facebook/relay/pull/2889

Test Plan: So far, I did a few test releases to https://www.npmjs.com/package/release-test-react-relay?activeTab=versions, but to test the final code, we'll have to push this.

Reviewed By: josephsavona, tyao1

Differential Revision: D17983155

Pulled By: kassens

fbshipit-source-id: 30dc3d69eb458584b734b98d528a92abf7b1ef33