Allow non-nullable Viewer field in ViewerQueryGenerator (#2924)
Summary:
This adds support for non-nullable `Viewer` field by using `getNullableType` before using the type of the viewer field.

Also make sure that when generating the refetch operation we use the actual type of the field (`viewerField.type` instead of `viewerType`) to ensure nullability matches the one in the schema.
Pull Request resolved: https://github.com/facebook/relay/pull/2924

Reviewed By: alunyov

Differential Revision: D18303605

Pulled By: kassens

fbshipit-source-id: 8ef2abdfd4a51b8778f9461db591e8a3869f15e4