Fallback to non watch mode instead of crashing (#2929)
Summary:
Changes the compiler to not error if `--watch` is passed and watchman is not installed. It's convenient for things like the example app to just be able to call the compiler with `--watch` and not crash if the user happens to not have watchman installed.
Pull Request resolved: https://github.com/facebook/relay/pull/2929

Test Plan: eyes

Reviewed By: alunyov

Differential Revision: D18336731

Pulled By: josephsavona

fbshipit-source-id: e7a1bc5f5f269b0f83e81553d2f8d514cc4dd42c