Fix: always check query in preloadQuery to account for staleness
Summary:
Currently we skip calling check inside preloadQuery_DEPRECATED if we already had a cache entry for that query.

However, that made it so we would reuse cached data if the cached entry hadn't been cleaned up, even if the underlying data was stale.

To fix this, this commit makes it so we always check for stale data.

Reviewed By: alunyov

Differential Revision: D19037720

fbshipit-source-id: 6405c839c3ca7f347d7040fdbb4f454c75d1fbba