Relay Compiler: setup external config conditionally
Relay Config package doesn't exist in every project which results in `yargs.config(undefined)` call. Unfortunately, yargs adds a new `--config` option for these cases. However, this option is not specified by Relay Compiler so it confuses people. I am not sure if it's a bug or feature though because technically it's fine as well. I'd argue that it's better to hide it and expose it manually (when needed) because now it's sometimes available and sometimes not.