Remove new connection model experiment
Summary: Removes support for the new connection model from across the compiler and runtime. This could be an interesting avenue to pursue in the future - and the exploration lead us to find a new approach to handling applying/reverting/rebasing optimistic updates - but overall we decided not to pursue this further right now.

Reviewed By: jstejada, kassens

Differential Revision: D19410941

fbshipit-source-id: 5917fb673c74112626f185675131d712728a0b93