RFC: Add unstable option to pass extra variables during pagination with hooks
Summary:
# Motivation
In very rare cases, it is necessary for product developers to pass variables to the pagination request that are different from the original initial query.

In such cases, these are variables that don't affect the output of the connection (i.e. they aren't filters), so they kind of just produce a side effect in the server, but which still need to be sent as with a different value for each request, e.g. a client request id.

There are other cases such as using the `find` filter on a connection to fetch a specific node in the connection. For this case, we don't want the pagination requests to  pass the original `find` value from the initial request.

# Changes

For this reason, this commit adds a new option to the `loadNext` function exposed by our pagination hooks: `UNSTABLE_extraVariables`.

- We use Flow to enforce that the extra variables match the variables expected by the pagination request, and added a flow test for that
- We ignore any pagination variables passed as extra variables.

# Caution

This should be used sparingly as it can cause unexpected results. If you fetch data that you are not rendering, e.g. change a filter, this wont update the connection that is being rendered.

Reviewed By: josephsavona

Differential Revision: D19412033

fbshipit-source-id: 1fb7918e5cb9ee34099236879ffd2fb0d0d068bd