Make .graphql extension be a part of the language plugin
This makes it possible to use the language plugin interface in java.

Background:

Relay will generate the file "MyFragment.graphql.java" and a class "MyFragment.graphql". This is however not valid java code, dots "." is not allow in class names.

It's possible to change the class name programatically to solve that problem, but then we will have a public class that do not match the name of the file, which as a result is going to create a compilation error.