Compiler.compiler returns a Result now
Summary:
The compiler execution engine now no longer produces console output to print errors and instead returns a `Result` object.

I had to attach the sources to the Error objects to allow them to print the source code around them. I'm not sure this is the best way, or maybe the location could have an Arc reference to the source code? I was thinking to go forward with this simple change for now.

Reviewed By: josephsavona

Differential Revision: D20103792

fbshipit-source-id: cfad4f1325d643d54d3982fce71335d7a5952714