Transformer::transform_value now returns TransformedValue
Summary:
Previously, `transform_value` would return a `Transformed` which is conceptually a member of a list which should be able to shrink (deleting items) or grow (adding items). This doesn't really make sense for a "value" used in places like a variable binding or argument. The *variable binding* can be removed, but not the value and bubble up to delete a the binding. This would be additionally complex when we want to talk about that `Transformed` (list element) could add additional items.

Also adds:
- `TransformedValue::map` which maps over replacements similar to `Transformed::map`.
- `impl<T> Into<Transformed<T>> for TransformedValue<T>` which allows a more restrictive TransformedValue to be converted into a `Transformed`. This allows us to use `transform_list` with functions that return `Transformed` as well as `TransformedValue`.

Reviewed By: josephsavona

Differential Revision: D20356436

fbshipit-source-id: a3f56ac92bedd205075c630100c49ea92b28e626