Add option in Environment to handle stripped nulls in payload (#3061)
Summary:
Adding the feature to handle stripped nulls in responses as an environment option, passing it through to the query executor and then to the response normalizer, which actually does the job of writing the fields with nulls or not based on the flag.

In the QueryExecutor, I explicitly passed the value to `false` when processing optimistic responses. I noticed that this path is not only used for optimistic responses defined in cases such as `commitMutation`, but also executed when the server may send an optimistic response using an extension. I think this case should still be stripping nulls, as those are coming from the server (if I understood correctly). Would love to confirm this and change the code so that I take into account that case (and change the test I added).

Also pass the value through for all the other async cases, such as defer, streams, module imports. I'm not very familiar with how those work and how they should play along with this option, so I'd appreciate some feedback. Based on [this comment](https://github.com/facebook/relay/issues/3052#issuecomment-602788832), it seems that based on how they are traversed it should not be an issue.

Finally, if this approach seems to be the way to go, happy to work on adding documentation for this (would love pointers on where would be the best place to put this).

Fixes https://github.com/facebook/relay/issues/3052
Pull Request resolved: https://github.com/facebook/relay/pull/3061

Reviewed By: josephsavona

Differential Revision: D21066513

Pulled By: jstejada

fbshipit-source-id: 86e34dc08cfcdacdc9b3679d847f3ddd1794240f