Guarantee that we generate type discriminators for all abstract type refinements
Summary:
allow-large-files

This commit updates our JS compiler to ensure that we /always/ generate a type discriminator in our codegen when we do an abstract type refinement.

## Context

We've done a couple of changes to abstract type refinement with the goal of more precisely determining when data is missing so that we don't incorrectly suspend when we shouldn't have or don't incorrectly refetch a query when we shouldn't have (T58655791)

In D21328939, we updated the compiler codegen to add type discriminators when we encountered an abstract type refinement to give us information at runtime about whether a record of a concrete type actually implements the type of the abstract refinement.

Using this information, in D21390159, we landed a fix for detecting missing data at runtime in which we check if the concrete type implements the abstract type for the abstract refinement, and only if so do we consider data to be missing.

However, as was pointed out in the summary of D21390159, there is a caveat: fragment //spreads// on abstract types won't always know if the type implements the interface or not. Consider:

```
query {
  node(id: "4") {
     ...NodeFragment
  }
}

fragment NodeFragment on Node {
  id
}
```

Given our current compiler pipeleine, the normalization node inlines the spread, and then flattens the InlineFragment since `node` returns `Node`, so this turns into the following for the normalization ast:

```
query {
  node(id: "4") {
    # oops no intermediate InlineFragment, no '__isNode' selection
    id
  }
}
```

So when we go to read `NodeFragment` it will appear as if the object doesn't implement `Node`, so we'll incorrectly reset `isMissingData`.

A similar thing may happen when we skip redundant nodes and end up with an empty inline fragment. For example, consider:

```
query {
  node(id: "4") {
    __typename
    id
    ... on Actor {
      actor_field
      ...NodeFragment
    }
  }
}

fragment NodeFragment on Node {
  id
}
```

In this case, after compilation, the fragment spread will be inlined, and the `id` field will be skipped since it's already fetched in the parent, leaving us with an empty inline fragment that gets removed:

```
query {
  node(id: "4") {
    # oops no intermediate InlineFragment, no '__isNode' selection
    id
    ... on Actor {
      actor_field
    }
  }
}
```

Here, again, when we go to read `NodeFragment` it will appear as if the object doesn't implement `Node`, so we'll incorrectly reset `isMissingData`.

As of D21390159, if we were to encounter this case where we can't accurately tell if an object implements the interface, we decided to tolerate the impreciseness and consider that case as **not** having missing data, since incorrectly considering the data as missing seemed like a worse outcome.

However, upon [rolling this out](https://fb.workplace.com/groups/225341071639862/permalink/653824078791557/) we discovered that not suspending in some cases where we should've caused runtime errors, so we had to roll back.

## Changes

In order to properly fix this, we need to take the original (most precise) approach of considering data to be missing if we can't accurately tell if a record implements an abstract type (i.e. if the type discriminator is missing).

But in order for this assumption to be safe, we need to **guarantee that we are always fetching the type discriminator and recording its type information**. In order to do this, we need to address the caveats mentioned above.

To fix this, this commit introduces the following changes to our compiler, as suggested by josephsavona:

- Move `GenerateTypeNameTransform` ***before*** flattening and skipping redundant nodes, and ***always*** generate the scalar field selection for the type discriminator (`__isFoo`) under abstract inline fragments and fragment definitions.
  - Having this transform run *first* will ensure that the type discriminator field selection is always present before flattening and skipping nodes, meaning that we will visit all abstract inline fragments before they are possibly flattened away or skipped. With the type discriminator always present, we can then dedupe and flattened the IR as normal during our codegen pipeline.
  - We skip generating the type discriminator for abstract types that are client-defined
- FlattenTransform and SkipRedundantNodesTransform don't need to be updated; since we are always generating the type discriminator in the previous step, the type discriminator will be treated as s regular selection and it can be deduped, and inline fragments might be flattened or skipped as well.
  - Note that while these steps may remove inline fragments that had an `abstractKey`, the type discriminator selection will never be fully removed; at most, it will be deduped.
- During codegen, we emit a new `TypeDiscriminator` runtime node, **only for normalization ast**:
  - Reader ast doesn't need this node because it's not a selection we will ever attempt to read; Reader only needs to lookup the type information we've already stored during normalization.
  - During normalization, the `abstractKey` in this node will be used in the same way that the `abstractKey` on abstract inline fragments, which is to record the appropriate type information (instead of actually normalizing that field).
  - During normalization codegen, we also skip any TypeDiscriminator nodes that are direct children of an abstract inline fragment, since this is redundant information; the `TypeDiscriminator` is only necessary for the cases when its parent abstract inline fragment had been flattened or skipped as a result of our transforms.

## Next steps

Once these changes land, we will guarantee that we always generate the type discriminator information in codegen where needed.

With this, we can:

- Update our runtime to properly consider the type discriminator missing as actual missing data that we would suspend on, and stop treating the case of "not knowing" as "not missing" (next diff in stack)
- Roll out the changes again with [`relay_enable_precise_type_refinement`](https://our.intern.facebook.com/intern/gatekeeper/projects/relay_enable_precise_type_refinement/) GK
- Remove lint rule that disallows abstract type refinement (T66676710)

(Note: this ignores all push blocking failures!)

Reviewed By: kassens

Differential Revision: D21541588

fbshipit-source-id: 4aba6ff9ed88dbfd268a110784b1efd7d54efd7e