skip_redundant: only memoize shared fields/fragments
Summary: Optimizes the memoization strategy for skip_redundant_nodes, *reducing* the amount of selections that are memoized, which can improve performance by keeping the memo cache smaller (less allocations) and reducing thread contention (DashMap is shareded, but still there can be contention on shards). Before we memoized *all* selections whenever the selection_map was empty, now we only memoize LinkedFields and InlineFragments, only when the selection map is empty, and *only when the frield/fragment is shared*. Non-shared selections have only one reference to them, so there's no benefit to caching them.

Reviewed By: kassens

Differential Revision: D22050877

fbshipit-source-id: ff7d1cbf9ec13360ec275d3b8df7eee10d6b2295