Support ScalarHandle in DataChecker (#3100)
Summary:
If we have a fragment that makes use of a handler in a scalar field, `DataChecker` will throw an error because the `SCALAR_HANDLE` kind of selection is not supported. Talking with the team we agreed this was a case that should be supported.

I did this by replicating the logic for LINKED_HANDLE, only that in this case we don't need to copy inner selections since this is a scalar field. I created a `cloneRelayScalarHandleSourceField` module similar to its linked handle counterpart. Also while at it, I added some tests to support this functionality and that of LINKED_HANDLE, since it was previously not covered.

### **Wrong assumptions based on kassens's [comment below,](https://github.com/facebook/relay/pull/3100#issuecomment-647920623) leaving this for posterity:**

~~When figuring out the logic around a handle when checking if the data is complete, given that a handle field lives alongside the original one (which contains the data that came back from the response), it turned out that it would be enough to just skip this selection, as the logic for the original field would be enough to test that the data is complete.~~

~~I went a bit further with this reasoning and did the same with a `LINKED_HANDLE`. The logic here was even creating a clone of the original field and assigning the sub selections, traversing that in the same way we do for linked fields. But if the original linked field is always present next to a handler, then this is already being done when traversing the original field, so there should be no need to do this.~~

~~Probably I'm missing something in the assumptions for the latter case, but figured I would just send a PR and explaining the rationale to start the discussion. However, worth noting that all tests still pass even when skipping a LINKED_HANDLE field. So if this is not the right thing to do, we should at least add a test that validates this is indeed used.~~

Pull Request resolved: https://github.com/facebook/relay/pull/3100

Reviewed By: kassens

Differential Revision: D22478180

Pulled By: josephsavona

fbshipit-source-id: d2a82ab967cbbe8fe2a371afa71e192ba6d01976