parent_type on Field
Summary: Adds an optional `parent_type` property on schema `Field` objects, which allows navigating to the type that defines the field. This is useful for transforms that want to rewrite calls of one field of a type into a different type, or optionally add calls of another field if it exists, etc. Note that this is distinct from the "parent type" that we often track during a traversal, as that is influenced by parent selections such as inline fragments.

Reviewed By: kassens

Differential Revision: D22904784

fbshipit-source-id: 4339ee5b62a5e0c4a8290464f39e3070b1a81fe5