move definition of SyntaxResult to syntax_error.rs
Summary:
This stack starts to move the schema parsing into the `graphql-syntax` crate. Eventually, they should share one lexer and parser, maybe with variants to parse "schema only documents". One example that we should get rid of the schema parser is that it's error handling is pretty bad and we should have better errors.

This diff moves `SyntaxResult` next to the `SyntaxError` it wraps and `ParseResult` into the only file that uses is and makes it private.

Reviewed By: josephsavona

Differential Revision: D22922961

fbshipit-source-id: cd3efa2d3858bbdbe66b3a0d5bd41f846ea092f2