rename syntax_node to executable_node
Summary:
This stack starts to move the schema parsing into the `graphql-syntax` crate. Eventually, they should share one lexer and parser, maybe with variants to parse "schema only documents". One example that we should get rid of the schema parser is that it's error handling is pretty bad and we should have better errors.

This diff just renames the `syntax_node`  module to `executable_node` to disambiguate from a future `type_system_node`.

Reviewed By: josephsavona

Differential Revision: D22922962

fbshipit-source-id: 0a8f6fc3065f6ec4acbd847905368adfbdddd13c