Add support for variable definitions syntax
Summary:
graphql-js already supports an experimental syntax allowing fragments to define the variables they reference:

```
fragment Foo($localId: ID!) on User {
  id
}
```

This diff adds support for this syntax, treating it equivalently to `argumentDefinitions`, and adds validation that the syntax cannot be combined with that directive. This syntax extension is disabled by default, just as in graphql-js.

Reviewed By: kassens

Differential Revision: D23214991

fbshipit-source-id: 78fe11c8f993ff87cfe1158290a4af99b2de923c