Upgrade fbjs to 2.0.0 (#3184)
Summary:
Every time I run `yarn upgrade` on a personal project it spits out a warning:
```
warning react-relay > fbjs > core-js@2.6.11: core-js@<3 is no longer maintained and not
recommended for usage due to the number of issues. Please, upgrade your dependencies
to the actual version of core-js@3.
```

[New versions of fbjs and fbjs-scripts](https://github.com/facebook/fbjs/commit/1f144964c07c834ab548e274fd8b102901ef00b3) were released back in July that included a PR to [upgrade `core-js` to version 3](https://github.com/facebook/fbjs/pull/370).

This bumps the `fbjs` and `fbjs-scripts` versions to 2.0.0 for relay and all of its packages, which eliminates `core-js@<3` completely!

Pull Request resolved: https://github.com/facebook/relay/pull/3184

Reviewed By: kassens

Differential Revision: D23667268

Pulled By: bigfootjon

fbshipit-source-id: f7ee028d6555b43af63b2e913e5db5ba78394a06