implement deleteEdge store updater directive (#3177)
Summary:
_This supersedes https://github.com/facebook/relay/pull/3148 which I've now closed._

This implements a new declarative for updating the store declaratively after mutations called `deleteEdge`. As the name hints at, this particular directive allows you to remove a node's edge from the provided connections. Please note that this directive does not _delete the node_, only edge(s) for the node. There's `deleteRecord` already for deleting a record, which can be combined with this directive.

It's intended to be used like this:

```
mutation DeleteComment($input: DeleteCommentInput!, $connections: [String!]!) {
  deleteComment(input: $input) {
    deletedCommentId deleteEdge(connections: $connections)
    # This will delete any edge for the node with id `deletedCommentId` from the
    # connections provided through `$connections`
  }
}
```

It works for single IDs (as demonstrated above) as well as a list of IDs.

Pull Request resolved: https://github.com/facebook/relay/pull/3177

Reviewed By: kassens

Differential Revision: D23652072

Pulled By: tyao1

fbshipit-source-id: 0d41dd0bb4950276bd8a8fb9178b1ee9d0a4337d