delete schema-diff for now
Summary:
This crate is the last place using the old schema AST. We should update it to the new AST before using it to ensure consistent parsing.

Since we don't have immediate plans for using this, I think we can delete this code for now to enable the next diff deleting the old parser.

When we want to use this in the future, reverting this diff and upgrading it to the new parser should be about as much work as it's now to upgrade.

Reviewed By: tyao1

Differential Revision: D24111458

fbshipit-source-id: c83086239423a41e8b7c4dee92dafd1ba5dac62c