change the ID type for directives handling connection IDs (#3202)
Summary:
This changes the type of the `connections: [String]` arguments to the new store updater directives handling connections from `String` to `ID`. This is to make it match better type-wise and semantically to what it actually is (an ID coming from the store), but since both `String` and `ID` normally serialize to `string`, it won't change the behavior of any product code.

It is however a breaking change in the sense that all operations that use the new store directives with `connections` will need to change from `String` to `ID` in the definition itself.

Pull Request resolved: https://github.com/facebook/relay/pull/3202

Reviewed By: jstejada

Differential Revision: D24219086

Pulled By: kassens

fbshipit-source-id: b3669f48601da5525bdb90773e22cb2e7b489dec