implements hasOwnProperty as a util
When running relay and using fragments at various points. (E.G.: RelayResponseNormalizer line 236, I would hit an issue where an object would be stripped of it's prototype causing hasOwnProperty to be undefined.) Refactoring to always use the hasOwnProperty on Object's prototype avoids these issues and should be safer as long as the root prototype is altered.