validate required arguments when building IR
Summary:
Previously, required arguments were not validated when constructing the IR. This was pretty counter to the other type level validations happening at that stage such as validating that arguments that are passed are of the correct type.

Unfortunately, Relay has some special cases.

1. The `label` argument to `defer` and `stream` is not required by Relay and generated if missing. Solution here: patch the schema in the `relay-schema` crate to make the argument nullable.
2. If a `match` directive is present on a linked field, Relay generates a `supported` argument and makes the argument not-required. The would require patching *all* fields in the schema, but wouldn't be able to capture if a field has the `match` directive. The somewhat ugly solution here is to add a `relay_mode` flag to build_ir that triggers the special case.

A longer term solution could be to make the IR builder an abstract trait and implement a `relay-ir-builder` crate that captures the special cases. I think this would allow us to extract these relay cases out of the non-Relay foundation.

Reviewed By: tyao1

Differential Revision: D24697562

fbshipit-source-id: 5690076dc53309a3aaad7963d66520260509723f