Fix retaining queries when re-mounting a query component
Summary:
This commit fixes the issue reported in https://github.com/facebook/relay/issues/3298

When re-mounting a query component without changing the query, before this commit we would briefly fully release the query and then immediately attempt to re-read it. If Relay GC is enabled and scheduled to run synchronously, when the query is fully released, we would synchronously delete the data and then immediately attempt to read the data from cache again when the query was remounting, only to find empty data.

The query was being fully released incorrectly because we didn't do our usual temporary retain during render if the query was already permanently retained; specifically, when the query is remounting, the **/next/** render will happen /before/ the current query is disposed/unmounted, so we ended up skipping the temporary retain for the new render.

This was incorrect since it leaves open the possibility for the data to be deleted before the render is able to commit, which is precisely the reason the temporary retain exists. In general, every single render should have a temporary retain associated with it, even if it never ends up committing.

The fix in this commit is to never skip the temporary retain. The issue is further illustrated in the regression test that was added.

Reviewed By: rbalicki2

Differential Revision: D26326454

fbshipit-source-id: b74449b95c6e8a55996f05fccb757029c52e6426