Initial, gated support for shared normalization artifacts
Summary:
## Summary

Relay normalization ASTs represent the transitive selections of an operation (query, mutation, subscription), including all reachable fragments. This is necessary so that Relay can understand the operation at runtime and do things like process query results, or check if a query can be fulfilled from the store, etc. Currently, *all* fragment spreads are inlined within a normalization AST - ie there are no FragmentSpread nodes in the normalization AST, only InlineFragments.

The motivation for this was (and is) that there is often significant overlap between the selections of different fragments. We want to encourage developers to colocate their data dependencies, with each component requesting just the data it needs, to maintain component encapsulation and allow developers to reason locally. Developers shouldn't have to worry about whether three sibling components happen to fetch the same field - Relay (Compiler) should optimize that for them. Inlining allows us to compile away such duplication and make the normalization smaller *in the common case*.

However, in any application *some* fragments may be spread many times across different queries and/or in many places within a given query. For such highly shared fragments, the costs of inlining (increasing the size of the ast at each usage site) may outweigh the benefits - ie we may not be able to de-duplicate enough fields at each call site to outweigh the duplication from inlining.

This diff introduces a new directive `no_inline` which tells the compiler not to inline that fragment within the normalization ast. Note that this has *no* impact on how data is read back out to components, or how data is processed at runtime, this is purely about how the normalization ast is represented in order to balance the benefits of deduplication from inlining with the desire for more structural sharing of normalization asts.

## Key Changes

This diff includes a few changes:
* Support for non-inlined fragments via the `no_inline` directive, gated by a feature flag which allows disabling/enabling the feature completely OR enabling it partially for a specific set of fragments.
* Support for rewriting fragment arguments to preserve fragment argument semantics even with shared normalization asts, w/o shadowing variables incorrectly. Most of the changes in this diff are in service of handling fragment-local variables correctly.

## How Fragment Arguments Works Today
The compiler first "applies" fragment arguments, effectively making a unique copy of each fragment for every combination of args that it is called with. Then the fragment spreads are inlined. Collectively, this ensures that _any_ variable usage within a fragment is replaced such that:
* It is replaced with a constant, if a constant was passed as an `arguments` value which flowed to that location
* OR is replaced with a query variable (ie, the name of a variable that is defined on the query that contains this now-inlined fragment).

The "apply arguments" step is more like abstract interpretation - it figures out which "values" (query variables or constants) flow into which local variables.

## Supporting Non-Inlined Fragments

Supporting non-inlined (stated conversely: shared) normalization fragments implies both not inlining //and// changing how we apply fragment arguments at the boundaries of non-inlined fragments.

One option is to add full runtime support in normalizer, datachecker, and referencemarker for fragment-local variables. Every fragment under a non-inline point would have to be made non-inlined, and every single spread would require re-mapping args to definitions. That's really inefficient.

Instead, what I think we want is to keep coarse-grained units of normalization nodes and //only// split at explicitly defined non-inline boundaries. Within those boundaries fragments would still be inlined (and args applied) as we do today. The challenge is how to distinguish whether a given variable within such an AST refers to a query global or whether it refers to a value that came from an argument to the non-inlined fragment:

```
query Foo($query_scale: Float) {
  me {
    # spread of a non-inlined fragment, so this spread stays in the normalization ast
    ...MeFragment args(const: 2.0, from_parent: $query_scale)
  }
}

fragment MeFragment on User
no_inline # <-- not inlined
argDefs(
  const: Float,
  from_parent: Float
) {
  # references a local: we need to map this to make sure it can't conflict w a query var of same name
  const: profile_picture(scale: $const) { uri }
  # references a local: we need to map this to make sure it can't conflict w a query var of same name
  from_parent: profile_picture(scale: $from_parent) { uri }
  # this refers to a query variable, since there's no local of this name
  global: profile_picture(scale: $query_scale) { uri }

  # passes a query var down: usages of that local must transform to reference the global
  # crucially: this is how apply fragment arguments transform works already!
  ...Child args(var: $query_scale)
}
```

The idea (implemented here) is to rename the locals of the non-inlined fragment ('foo' => 'FragmentName$foo'). We then apply fragment arguments within the non-inlined boundary normally. Any spread that references the non-inlined fragment has to also rename its arguments to match the updated variable names.

At runtime, we would keep two sets of variables available: the globals and the current scope. - initially the same object. Whenever we encounter a spread, we copy the globals and then merge in the values of the spreads arguments (which are already renamed to match the locals). This becomes the current scope while we're within that spread, then when we exit we put back the previous scope.

This keeps the runtime simple (all variable lookup is to a single object - we don't need to distinguish "local" vs. "global" variables at runtime. It also allows us to continue inlining constants that are passed on regular inlined spreads.

## Checklist

Note: "[ ]" is todo, "[x]" is done.

- [x] Update apply_fragment_arguments():
  - [x] Rewrite fragment spread args per the above scheme
  - [x] Convert non-inlined (shared) fragments into SplitOperation nodes (Operation w a special directive)
    - [x] SplitOperation name is `<FragmentName>$normalization`; fragment spread name is updated to match
- [x] inline_fragments() ignores spreads of non-inlined fragments
- [x] skip_redundant_nodes() ignores spreads of non-inlined fragments
- [x] codegen outputs argDefs for SplitOperation nodes
- [x] codegen outputs normalization FragmentSpread as a require call of the fragment normalization file
- [x] Test on www codebase
- [x] relay_test_operation support
- [=] raw_resonse_tpe support
  - NOTE: for now, we simply exclude the selections from the non-inlined fragment
- [x] Enable conditionally based a per-project compiler feature flag, w support disabled by default
  - There are 3 options: fully disabled, fully enabled, and a "limited" mode that allows only a specific set of fragments by name (and where fragments w variable definitions are disallowed)

For a follow-up diff (ok since the feature is off by default):
- [ ] Update runtime; runtime tests
  - [ ] Update NormalizationNode flow type to reflect new variant
  - [ ] Normalizer
  - [ ] DataChecker
  - [ ] ReferenceMarker (GC)

Longer-term:
- [ ] Test and enable for fragments w variables
- [ ] Ensure this works with `refetchable` and RefetchContainer
- [ ] Integrate with `module` to allowing passing `arguments` to 3D fragments - the 3D fragment would be treated as a shared normalization fragment, subject to the same variable renaming.

Reviewed By: kassens

Differential Revision: D26244754

fbshipit-source-id: 3ea80cebaae6356d66edcfaed34598a51bf43a3d