Use getRawType to unwrap any wrappers around edge type before getting fields (#3340)
Summary:
Currently, if you try to use `prependEdge` or `appendEdge` on a field whose type is non-null, you get this error when running the compiler:

```
ERROR:
Internal Error: _getFieldsMap: Type 'undefined' should have fields.
```

This happens because the type is wrapped. There was already unwrapping logic for handling lists of edges, but replacing that with `getRawType` should handle any amount of list and non-null wrapping.

Pull Request resolved: https://github.com/facebook/relay/pull/3340

Reviewed By: jstejada

Differential Revision: D26487172

Pulled By: josephsavona

fbshipit-source-id: 1827a3fcd8664dfb07df01ab01eee0e9bfe182fd