Configurable module import formats
Summary: The compiler emits some `require()` statements, but the format for this is currently hard-coded to `require('<AbsoluteModuleId>')`. This diff makes the format configurable via the project config, with two options for now: Haste (same behavior as today) and CommonJS (uses relative requires). I expect we'll iterate on this, but this at least sets up the scaffolding for adding more options / further customization, and unblocks rolling out `no_inline`.

Reviewed By: kassens

Differential Revision: D26497308

fbshipit-source-id: 53a6e27eb27dc32d658ae2ca256f21fb579e3615