disallow unused @argumentDefinitions
Summary:
This adds validation for unused local `argumentDefinitions` which are often a sign of a typo or misunderstanding them as global variables.

`argumentDefinitions` can now also be tagged as `unusedLocalVariable_DEPRECATED: true` to allow (actually requires) them to be unused. If the argument starts being used, the suppression has to be removed.

This also comes with a config setting `skipUnusedFragmentVariableValidation` which can disable the validation. Even when disabling, `unusedLocalVariable_DEPRECATED` can already be added. This allows incremental additions of the suppression before enabling the validation.

Reviewed By: josephsavona

Differential Revision: D27160302

fbshipit-source-id: 848d61d9585795db4238487fd2890212acd267af