A better intro/description to the usePreloadedQuery hook
Instead of straight away asking the reader to go through the code and then understand what the hook does, the committed line gives a better intro of what the hook does. Also, what the inputs to the hook are and what it's output is. I feel like this important detail should be highlighted better.