A better intro/description to the usePreloadedQuery hook (#3453)
Summary:
Instead of straight away asking the reader to go through the code and then understand what the hook does, the committed line gives a better intro of what the hook does. Also, what the inputs to the hook are and what it's output is. I feel like this important detail should be highlighted better. It gives a better context before reading the code

Pull Request resolved: https://github.com/facebook/relay/pull/3453

Reviewed By: josephsavona

Differential Revision: D27764874

Pulled By: kassens

fbshipit-source-id: a34aae1f403d5166f69b27c156758b3c52e59f07