Remove misleading advice regarding single artifact directory
The advice regarding the configuration of a single artifact directory to enable typing of fragment references appears to be outdated. Currently fragment references are typed appropriately when imported from a co-located `__generated__` path. The background information about Haste is interesting, but given this feature is no longer _required_, I suggest removing it here to avoid confusion.