fix incorrect parent_documents in relay_client_component transform
Summary:
The `parent_documents` should be the set of documents that are "cause" of a $normalization file to exist. This is so we can remove the generated files if all those sources were deleted.

This property wasn't set correctly, which caused some cases of incremental builds to incorrectly delete the $normalization files.

The previous diff in this stack updates the documentation and naming a bit to try to make this cleaner.

This diff fixes the issue in the relay_client_component and no_inline transforms.

Reviewed By: josephsavona

Differential Revision: D27909411

fbshipit-source-id: cf2e9ac5732c932716e15b40e7131c13f3d013fa